Comit about week 4
In this week we starting to use 'dbt' , as a tool to modeling or data in our DW.
With this tool, we can use all processing power from our DW and not from our local PC / Server.
And, in the end of the day,  all Transformations are done by SQL in DW engine.